requirements.txt
python>=3.12.9
babel==2.17.0
backrefs==5.9
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
ghp-import==2.1.0
idna==3.10
Jinja2==3.1.6
Markdown==3.8.2
MarkupSafe==3.0.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.6.16
mkdocs-material-extensions==1.3.1
packaging==25.0
paginate==0.5.7
pathspec==0.12.1
platformdirs==4.3.8
Pygments==2.19.2
pymdown-extensions==10.16.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
pyyaml_env_tag==1.1
requests==2.32.4
six==1.17.0
urllib3==2.5.0
watchdog==6.0.0
